Freeze no-arg instances by default, override with memo: false
If an instance method of a sum type with a no-arg variant instance
memoizes a result (or sets an instance variable for any other reason),
that will fail because the variant is frozen. This can be overridden
by passing `memo: false` into the variant initializer.

I'm making this the default to encourage stateless instance
methods (if any) in the defined types. Returning a memoized variant
instance without freezing it is dangerous because instance variables
set on that instance would live for the entire process, which could
lead to difficult to track down state related bugs.